+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog].
 ## [Unreleased]
 ### Added
 - Automatic bounds computation in MergeSkinnedMesh `#264`
+- Automatic Configuration System `#265`
+  - Currently FreezeBlendShape can be automatically configured.
 
 ### Changed
 - Support newly activated avatars in play mode for apply on play `#263`